// apps/admin/src/lib/api/generated/models/graphOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.1
 */
import type { GraphEdge } from './graphEdge';
import type { GraphNode } from './graphNode';
import type { GraphOutCatalogCounts } from './graphOutCatalogCounts';

export interface GraphOut {
  catalog_counts: GraphOutCatalogCounts;
  edge_limit: number;
  edges: GraphEdge[];
  node_limit: number;
  nodes: GraphNode[];
  truncated: boolean;
}